Backend/requirements.txt
langchain>=0.1.0
langchain-community
langchain-huggingface
faiss-cpu
huggingface-hub
google-generativeai
PyPDF2
tqdm
pypdf
sentence-transformers
python-dotenv
